requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
python-dotenv==0.20.0
mysql-connector-python==8.0.29
requests==2.27.1
lxml==4.8.0
